containerize & refactor
