+++ src/config.py
@@ -1,5 +1,3 @@
-API_HOSTNAME = "127.0.0.1"
-API_PORT = 1470
 LOGS_DIR = "logs"
 VERBOSE = True
 TIMESTAMP_FORMAT = "%Y-%m-%d %H:%M:%S"
